Changed from @SessionAttributes to HttpSession.get/setAttribute to handle session attributes since @SessionAttributes is not for permanent session attributes and behaves very strangely when having concurrent sessions it seems.
Implemented a way to reload pages (the reload itself can be changed to something else which is appropriate) triggered by the update of the patientContext in the same session. For now this is made with a long polling request. We'll see/discuss whether this is ok with regard to scalability.